run/realh-capital-evening-can-slim-scout-job/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-genai
google-auth
requests